llmgamemaster/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
openai==0.28.0
anthropic==0.34.0
jinja2==3.1.2
pyjwt==2.7.0
bcrypt==4.0.1
aiohttp==3.9.1
python-multipart==0.0.6